% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_reads.R
\name{count_reads}
\alias{count_reads}
\title{Count the number of reads per bin from a file}
\usage{
count_reads(
  bamfile,
  binsize = "10000",
  genome = "hg19",
  minMapq = 20,
  isPaired = TRUE,
  isProperPair = NA,
  isUnmappedQuery = FALSE,
  hasUnmappedMate = NA,
  isMinusStrand = NA,
  isMateMinusStrand = NA,
  isFirstMateRead = NA,
  isSecondMateRead = NA,
  isSecondaryAlignment = NA,
  isNotPassingQualityControls = FALSE,
  isDuplicate = FALSE
)
}
\arguments{
\item{bamfile}{Absolute path to a bamfile}

\item{binsize}{binsize to use, default = "10000", use quotes!}

\item{genome}{Reference genome (hg19 or hg38), default = "hg19"}

\item{minMapq}{filter reads with mapq values below this, default = 20}

\item{isDuplicate}{}
}
\value{
Data.frame with read counts and gc and mappability per bin
}
\description{
Count the number of reads per bin from a file
}
